backend/requirements.txt
starlette
python-multipart
uvicorn
databases[sqlite]
python-dotenv